+++ ISL.ReIdentification.Core.Tests.Unit/Services/Orchestrations/Accesses/AccessOrchestrationServiceTests.Logic.ValidateAccessForIdentificationRequests.cs
@@ -14,7 +14,7 @@ namespace ISL.ReIdentification.Core.Tests.Unit.Services.Orchestrations.Accesses
     public partial class AccessOrchestrationServiceTests
     {
         [Fact]
-        public async Task ShouldSetHasAccessToTrueIfUserHasAccessValidateAccessForIdentificationRequests()
+        public async Task ShouldValidateAccessForIdentificationRequests()
         {
             // given
             var accessOrchestrationServiceMock = new Mock<AccessOrchestrationService>
@@ -67,58 +67,58 @@ namespace ISL.ReIdentification.Core.Tests.Unit.Services.Orchestrations.Accesses
             accessOrchestrationServiceMock.VerifyNoOtherCalls();
         }
 
-        [Fact]
-        public async Task ShouldSetHasAccessToFalseIfUserDoesNotHaveAccessValidateAccessForIdentificationRequests()
-        {
-            // given
-            var accessOrchestrationServiceMock = new Mock<AccessOrchestrationService>
-                (this.dateTimeBrokerMock.Object,
-                this.userAccessServiceMock.Object,
-                this.pdsDataServiceMock.Object,
-                this.loggingBrokerMock.Object)
-            { CallBase = true };
+        //[Fact]
+        //public async Task ShouldSetHasAccessToFalseIfUserDoesNotHaveAccessValidateAccessForIdentificationRequests()
+        //{
+        //    // given
+        //    var accessOrchestrationServiceMock = new Mock<AccessOrchestrationService>
+        //        (this.dateTimeBrokerMock.Object,
+        //        this.userAccessServiceMock.Object,
+        //        this.pdsDataServiceMock.Object,
+        //        this.loggingBrokerMock.Object)
+        //    { CallBase = true };
 
-            string userEmail = GetRandomStringWithLength(10);
-            string inputUserEmail = userEmail;
-            string identifier = GetRandomStringWithLength(5);
-            string inputIdentifier = identifier;
-            AccessRequest accessRequest = CreateRandomAccessRequest();
-            accessRequest.IdentificationRequest.UserIdentifier = userEmail;
-            accessRequest.IdentificationRequest.IdentificationItems[0].Identifier = inputIdentifier;
-            string userOrganisation = GetRandomStringWithLength(5);
+        //    string userEmail = GetRandomStringWithLength(10);
+        //    string inputUserEmail = userEmail;
+        //    string identifier = GetRandomStringWithLength(5);
+        //    string inputIdentifier = identifier;
+        //    AccessRequest accessRequest = CreateRandomAccessRequest();
+        //    accessRequest.IdentificationRequest.UserIdentifier = userEmail;
+        //    accessRequest.IdentificationRequest.IdentificationItems[0].Identifier = inputIdentifier;
+        //    string userOrganisation = GetRandomStringWithLength(5);
 
-            List<string> userOrganisations =
-                new List<string> { userOrganisation };
+        //    List<string> userOrganisations =
+        //        new List<string> { userOrganisation };
 
-            bool userHasAccessToPatientResult = false;
+        //    bool userHasAccessToPatientResult = false;
 
-            accessOrchestrationServiceMock.Setup(service =>
-                service.GetOrganisationsForUserAsync(inputUserEmail))
-                    .ReturnsAsync(userOrganisations);
+        //    accessOrchestrationServiceMock.Setup(service =>
+        //        service.GetOrganisationsForUserAsync(inputUserEmail))
+        //            .ReturnsAsync(userOrganisations);
 
-            accessOrchestrationServiceMock.Setup(service =>
-                service.UserHasAccessToPatientAsync(inputIdentifier, userOrganisations))
-                    .ReturnsAsync(userHasAccessToPatientResult);
+        //    accessOrchestrationServiceMock.Setup(service =>
+        //        service.UserHasAccessToPatientAsync(inputIdentifier, userOrganisations))
+        //            .ReturnsAsync(userHasAccessToPatientResult);
 
-            AccessOrchestrationService accessOrchestrationService = accessOrchestrationServiceMock.Object;
-            AccessRequest expectedAccessRequest = accessRequest;
+        //    AccessOrchestrationService accessOrchestrationService = accessOrchestrationServiceMock.Object;
+        //    AccessRequest expectedAccessRequest = accessRequest;
 
-            // when
-            AccessRequest actualAccessRequest =
-                await accessOrchestrationService.ValidateAccessForIdentificationRequestAsync(accessRequest);
+        //    // when
+        //    AccessRequest actualAccessRequest =
+        //        await accessOrchestrationService.ValidateAccessForIdentificationRequestAsync(accessRequest);
 
-            // then
-            actualAccessRequest.Should().Be(expectedAccessRequest);
+        //    // then
+        //    actualAccessRequest.Should().Be(expectedAccessRequest);
 
-            accessOrchestrationServiceMock.Verify(service =>
-                service.GetOrganisationsForUserAsync(inputUserEmail),
-                    Times.Once);
+        //    accessOrchestrationServiceMock.Verify(service =>
+        //        service.GetOrganisationsForUserAsync(inputUserEmail),
+        //            Times.Once);
 
-            accessOrchestrationServiceMock.Verify(service =>
-                service.UserHasAccessToPatientAsync(inputIdentifier, userOrganisations),
-                    Times.Once);
+        //    accessOrchestrationServiceMock.Verify(service =>
+        //        service.UserHasAccessToPatientAsync(inputIdentifier, userOrganisations),
+        //            Times.Once);
 
-            accessOrchestrationServiceMock.VerifyNoOtherCalls();
-        }
+        //    accessOrchestrationServiceMock.VerifyNoOtherCalls();
+        //}
     }
 }
